requirements.txt
ultralytics
opencv-python
pytesseract
Pillow
ffmpeg-python